+++ packages/polaris-test/nest-server-schema-first/generate-typings.ts
@@ -6,5 +6,8 @@ definitionsFactory.generate({
     typePaths: ['./**/*.graphql'],
     path: join(process.cwd(), 'graphql.ts'),
     emitTypenameField: true,
+    customScalarTypeMapping: {
+        BigInt: 'BigInt',
+    },
     outputAs: 'class',
 });