+++ packages/polaris-test/nest-server-schema-first/graphql/resolvers/book.resolver.ts
@@ -116,6 +116,11 @@ export class BookResolver {
     ) {
         return this.bookService.createBook(title, authorId);
     }
+    @Mutation(() => Boolean)
+    public async createManyBooksSimultaneously() {
+        await Promise.all([this.bookService.createManyBooks(), this.bookService.createManyBooks()]);
+        return true;
+    }
     @Subscription()
     public bookUpdated() {
         return this.bookService.registerToBookUpdates();