+++ packages/polaris-test/server-without-connection/schema/type-defs.ts
@@ -14,6 +14,7 @@ export const typeDefs = `
         lastUpdatedBy: String
         lastUpdateTime: DateTime
         realityId: Int!
+        dataVersion: BigInt!
         title: String
         coverColor: String @upper
     }