+++ packages/polaris-test/server/schema/type-defs.ts
@@ -16,8 +16,9 @@ export const typeDefs = `
         permissionsFieldWithHeader: String @permissions(entityTypes: ["bar"], actions: ["READ", "DELETE"])
         onlinePaginatedBooks(pagingArgs: OnlinePagingInput!): BookConnection
         bookByDate(filter: EntityFilter): [Book]!
-        onlinePaginatedAuthors: [Author]!
+        onlinePaginatedAuthorsWithLeftJoin: [Author]!
         isThereTransactionActive: Boolean!
+        onlinePaginatedAuthorsWithInnerJoin: [Author]!
     }
 
     input ReviewKind{
@@ -31,11 +32,14 @@ export const typeDefs = `
         createPen(color: String!, id: String): Pen!
         createChapter(number: Int!, bookId: String): Chapter!
         createReview(description:String!, rating:String!, bookId: String!, reviewKind: ReviewKind!): Review!
+        createGenre(name: String!, bookId: String): Genre!
+        createOneToOneEntity(name: String!, bookId: String, genreId: String): OneToOneEntity!
         updateBooksByTitle(title: String!, newTitle: String!): [Book]!
         deleteBook(id: String!): Boolean
         deleteAuthor(id: String!): Boolean
         fail: Boolean
         createManyAuthors: Boolean!
+        createManyBooksSimultaneously: Boolean!
     }
 
     type Subscription {
@@ -50,10 +54,13 @@ export const typeDefs = `
         lastUpdatedBy: String
         lastUpdateTime: DateTime
         realityId: Int!
+        dataVersion: BigInt!
         title: String
         author: Author
         chapters: [Chapter]
         reviews: [Review]
+        genres: [Genre]
+        oneToOneEntity: OneToOneEntity
     }
     
     interface Review implements RepositoryEntity {
@@ -64,6 +71,7 @@ export const typeDefs = `
         lastUpdatedBy: String
         lastUpdateTime: DateTime
         realityId: Int!
+        dataVersion: BigInt!
         rating: Int!
         description: String!
         book: Book!
@@ -77,6 +85,7 @@ export const typeDefs = `
         lastUpdatedBy: String
         lastUpdateTime: DateTime
         realityId: Int!
+        dataVersion: BigInt!
         rating: Int!
         description: String!
         book: Book!
@@ -91,10 +100,39 @@ export const typeDefs = `
         lastUpdatedBy: String
         lastUpdateTime: DateTime
         realityId: Int!
+        dataVersion: BigInt!
         rating: Int!
         description: String!
         book: Book!
         name: String!
+    }
+    
+    type Genre implements RepositoryEntity {
+        id: String!
+        deleted: Boolean!
+        createdBy: String!
+        creationTime: DateTime!
+        lastUpdatedBy: String
+        lastUpdateTime: DateTime
+        realityId: Int!
+        dataVersion: BigInt!
+        name: String!
+        books: [Book]
+        oneToOneEntity: OneToOneEntity
+    }
+    
+    type OneToOneEntity implements RepositoryEntity {
+        id: String!
+        deleted: Boolean!
+        createdBy: String!
+        creationTime: DateTime!
+        lastUpdatedBy: String
+        lastUpdateTime: DateTime
+        realityId: Int!
+        dataVersion: BigInt!
+        name: String!
+        book: Book
+        genre: Genre
     }
 
     type Pen implements RepositoryEntity {
@@ -105,6 +143,7 @@ export const typeDefs = `
         lastUpdatedBy: String
         lastUpdateTime: DateTime
         realityId: Int!
+        dataVersion: BigInt!
         color: String
         author: Author
     }
@@ -117,6 +156,7 @@ export const typeDefs = `
         lastUpdatedBy: String
         lastUpdateTime: DateTime
         realityId: Int!
+        dataVersion: BigInt!
         number: Int!
         book: Book
     }
@@ -129,10 +169,13 @@ export const typeDefs = `
         lastUpdatedBy: String
         lastUpdateTime: DateTime
         realityId: Int!
+        dataVersion: BigInt!
         firstName: String
         lastName: String
         books: [Book]
         pens: [Pen]
+        country: String @deprecated
+        deprecatedField: String @deprecated(reason: "Will be removed in the next version")
     }
     
     type BookEdge {