+++ packages/polaris-test/tests/data-version.test.ts
@@ -60,4 +60,16 @@ describe('data version tests', () => {
             },
         );
     });
+    describe('parallel data version queries', () => {
+        test.each(createServers())(
+            'query with promise all, data version lock does not throw exception',
+            async (server) => {
+                await polarisTest(server, async () => {
+                    await graphQLRequest('mutation { createManyBooksSimultaneously }', {});
+                    const response = await graphQLRequest(allBooks.request, {});
+                    expect(response.allBooks.length).toBe(200);
+                });
+            },
+        );
+    });
 });