+++ packages/polaris-typeorm/src/typeorm-bypasses/polaris-repository.ts
@@ -103,10 +103,26 @@ export class PolarisRepository<Entity extends ObjectLiteral> extends Repository<
      * Used for online paging.
      * Finds entities and sorts them by the given data version(including their sub-entities)
      */
-    public findSortedByDataVersion(
+    public findSortedByDataVersionUsingLeftOuterJoin(
         optionsOrConditions?: FindManyOptions<Entity> | FindConditions<Entity>,
     ): Promise<Entity[]> {
-        return ((this.manager as unknown) as PolarisEntityManager).findSortedByDataVersion(
+        return ((this
+            .manager as unknown) as PolarisEntityManager).findSortedByDataVersionUsingLeftOuterJoin(
+            this.metadata.target as any,
+            optionsOrConditions,
+        );
+    }
+
+    /**
+     * Used for online paging.
+     * Finds entities and sorts them by the given data version(including their sub-entities)
+     * implemented by inner join and union
+     */
+    public findSortedByDataVersionUsingInnerJoin(
+        optionsOrConditions?: FindManyOptions<Entity> | FindConditions<Entity>,
+    ): Promise<Entity[]> {
+        return ((this
+            .manager as unknown) as PolarisEntityManager).findSortedByDataVersionUsingInnerJoin(
             this.metadata.target as any,
             optionsOrConditions,
         );