+++ packages/polaris-typeorm/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.1](https://github.com/Enigmatis/polaris-typeorm/compare/@enigmatis/polaris-typeorm@2.11.0...@enigmatis/polaris-typeorm@2.11.1) (2021-03-15)
+
+**Note:** Version bump only for package @enigmatis/polaris-typeorm
+
+
+
+
+
+# [2.11.0](https://github.com/Enigmatis/polaris-typeorm/compare/@enigmatis/polaris-typeorm@2.10.1...@enigmatis/polaris-typeorm@2.11.0) (2021-03-14)
+
+
+### Bug Fixes
+
+* dv handler and common model subscriber fixes ([2a80741](https://github.com/Enigmatis/polaris-typeorm/commit/2a80741ded4f1ac75f19b46a3b9a8c585c7a5e99))
+
+
+### Features
+
+* inner join implementation of online paging ([#70](https://github.com/Enigmatis/polaris-typeorm/issues/70)) ([43d110a](https://github.com/Enigmatis/polaris-typeorm/commit/43d110aac0ba1454aa929c6f9ff6e5582cac25ea))
+
+
+
+
+
+## [2.10.1](https://github.com/Enigmatis/polaris-typeorm/compare/@enigmatis/polaris-typeorm@2.10.0...@enigmatis/polaris-typeorm@2.10.1) (2021-03-10)
+
+
+### Bug Fixes
+
+* repo entities fields are now set when cascade save ([#69](https://github.com/Enigmatis/polaris-typeorm/issues/69)) ([bc6f401](https://github.com/Enigmatis/polaris-typeorm/commit/bc6f4013d263ffe750d0b0ba2a1f5a35bbdcb2e5))
+
+
+
+
+
 # [2.10.0](https://github.com/Enigmatis/polaris-typeorm/compare/@enigmatis/polaris-typeorm@2.9.0...@enigmatis/polaris-typeorm@2.10.0) (2021-03-07)